user soft delete implemented

+++ src/main/java/cooperate/app/business/user/login/LoginDto.java
@@ -7,7 +7,7 @@ import cooperate.infrastructure.mediation.IResponse;
 public class LoginDto implements IResponse {
     @NotMapped
     public boolean IsSuccess;
-    public long UserId;
+    public int UserId;
     public String Fullname;
     public byte[] Password;
     public int RoleId;
